Add documentation categories
This adds a flat category hierarchie for documentation files found
in the `_docs` folder. The list (and sort order) of categories is
defined as `docs_categories` in the `_config.yml`. Files without
a category (or an invalid category) are put into the default
category at the top of the navigational menu.

This commit additionally sets some default attributes for doc
files, such as the use of the `docs` layout and default
categories for files in certain subfolders.

Signed-off-by: Sebastian Wicki <swicki@inf.ethz.ch>